+++ Changelog.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - bugfix/28-cannot-auth-tools-list (2024-11-30)
 
 ### Changed 
+  - feature/35-switch-pip-for-uv (2024-12-01)
+  - feature/33-deploy-app-using-terraform (2024-12-01)
   - feature/30-tf-do-deployment (2024-11-30)
   - feature/25-add-basic-auth (2024-11-29)
   - feature/19-deploy-to-dev (2024-11-28)